+++ application/modules/Members/controllers/Members.php
@@ -2,24 +2,41 @@
 defined('BASEPATH') OR exit('No direct script access allowed');
 
 class Members extends CI_Controller {
+	
+		public function __construct()
+		{
+			parent::__construct();
+			//Load Dependencies
+	
+		}
+	
+		// List all your items
+		public function index( $offset = 0 )
+		{
+		
+		}
+	
+		// Add a new item
+		public function add()
+		{
+			$this->load->view('members_add');
+		}
 
-	/**
-	 * Index Page for this controller.
-	 *
-	 * Maps to the following URL
-	 * 		http://example.com/index.php/welcome
-	 *	- or -
-	 * 		http://example.com/index.php/welcome/index
-	 *	- or -
-	 * Since this controller is set as the default controller in
-	 * config/routes.php, it's displayed at http://example.com/
-	 *
-	 * So any other public methods not prefixed with an underscore will
-	 * map to /index.php/welcome/<method_name>
-	 * @see https://codeigniter.com/user_guide/general/urls.html
-	 */
-	public function index()
-	{
-		$this->load->view('members');
-	}
+		// Add a new item
+		public function store()
+		{
+			$this->form_validation->set_rules('title', 'Title', 'required');
+		}
+	
+		//Update one item
+		public function update( $id = NULL )
+		{
+	
+		}
+	
+		//Delete one item
+		public function delete( $id = NULL )
+		{
+	
+		}
 }